refactor(provider): improve overflow safety, NaN handling, and proxy error logging
- Use `u64` arithmetic in agent runner instead of truncating to `u32`
- Tighten quota/exceeded detection in error classifier (require both keywords)
- Add saturating multiplication for freeze backoff duration
- Handle NaN/Inf in `safe_f64_to_u64` with proper clamping
- Use direct `u64 as f64` conversions instead of detour through `u32`
- Log proxy creation failures instead of silently ignoring
- Use destructuring for ChatResponse fields to avoid clones

+++ crates/y-provider/src/agent_runner.rs
@@ -122,7 +122,8 @@ impl AgentRunner for SingleTurnRunner {
             });
         }
 
-        let tokens_used = response.usage.input_tokens + response.usage.output_tokens;
+        let tokens_used =
+            u64::from(response.usage.input_tokens) + u64::from(response.usage.output_tokens);
 
         Ok(AgentRunOutput {
             text,
